Add background music, quiz history; themed Voxspell.
> Makefile:
    * Compile recipe copies across stylesheets.
+ resources/Mars_Landscape_Sun_Set.mp3:
    * Background sound for main menu.
+ src/voxspell/gui/HistoryMenu.fxml:
    * Menu for viewing quiz results.
+ src/voxspell/gui/HistoryMenuController.java:
    * Added functionality to view past quizzes.
> src/voxspell/gui/MainMenu.fxml:
    * Moved components around.
    * Added button for viewing quiz history.
    * Fixed tooltip for logout button.
> src/voxspell/gui/MainMenuController.java:
    * Added functionality to access history menu.
    * Added functionality for background music.
> src/voxspell/gui/ResultMenu.fxml:
    * Moved VBox across for alignment.
> src/voxspell/gui/SceneSwitcher.java:
    * Added ability to change to the history menu.
    * Loads in stylesheet for the application before showing the
        scene when switching.
> src/voxspell/gui/SceneType.java:
    * Added enumeration for history menu.
> src/voxspell/gui/SettingsController.java:
    * Added confirmation dialogs for deleting user and clearing
        statistics.
> src/voxspell/gui/SpellingTest.fxml:
    * Removed results button.
> src/voxspell/gui/SpellingTestController.java:
    * Merged functionality of back to menu and view results into one
        button.
    * Only complete quizzes are added to the quiz history of the user.
> src/voxspell/gui/UserMenu.fxml:
    * Fixed issue with not being able to see all of the text for the
        CheckBox.
+ src/voxspell/gui/styles/main.css:
    * Implement theme for Voxspell.
+ src/voxspell/music/Music.java:
    * Singleton class that is responsible for playing and pausing
        the background music.
> src/voxspell/properties/lang_en.properties:
    * Added string for quiz history button.
> src/voxspell/spell/Quiz.java:
    * Added getter for the date/time that a quiz was taken.

+++ src/voxspell/gui/SettingsController.java
@@ -3,6 +3,7 @@ package voxspell.gui;
 import voxspell.user.User;
 import voxspell.spell.Topic;
 import voxspell.festival.Festival;
+import voxspell.music.Music;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -16,6 +17,8 @@ import javafx.scene.control.Label;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Spinner;
 import javafx.scene.control.SpinnerValueFactory;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 
 import javafx.scene.layout.AnchorPane;
 
@@ -59,16 +62,27 @@ public class SettingsController implements Initializable {
     @FXML
     void btnDeleteUserPressed(ActionEvent event) {
         //TODO: Show a confirmation dialog of some sort maybe?
-        User.deleteCurrentUser();
-        SceneSwitcher.getInstance().execute(SceneType.TITLE_MENU);
+        Alert confirmationDialog = new Alert(Alert.AlertType.CONFIRMATION);
+        confirmationDialog.setTitle("Delete User");
+        confirmationDialog.setContentText("Are you sure you want to delete this user (" + User.getInstance().getName() + ")? You can not undo this.");
+        if (confirmationDialog.showAndWait().get() == ButtonType.OK) {
+            User.deleteCurrentUser();
+            Music.getInstance().stop();
+            SceneSwitcher.getInstance().execute(SceneType.TITLE_MENU);
+        }
     }
 
     @FXML
     void btnClearStatsPressed(ActionEvent event) {
         //TODO: Show a confirmation dialog of some sort maybe?
-        User.getInstance().getSelectedTopicSet().clear();
-        User.getInstance().setTopicLevel(0);
-        User.getInstance().getQuizzes().clear();
+        Alert confirmationDialog = new Alert(Alert.AlertType.CONFIRMATION);
+        confirmationDialog.setTitle("Clear Statistics");
+        confirmationDialog.setContentText("Are you sure you want to clear the statistics for the current word list? You can not undo this.");
+        if (confirmationDialog.showAndWait().get() == ButtonType.OK) {
+            User.getInstance().getSelectedTopicSet().clear();
+            User.getInstance().setTopicLevel(0);
+            User.getInstance().getQuizzes().clear();
+        }
     }
 
     @FXML
